PMM-3578 update deps
