+++ Gopkg.lock
@@ -63,10 +63,11 @@
   revision = "aa810b61a9c79d51363740d207bb46cf8e620ed5"
 
 [[projects]]
-  digest = "1:69b5bdc0545fa057e73e7704cb9f7e8c48b9bb6f7cbe2b25686c0a67bba52276"
+  digest = "1:49c6e6f1a17423b3a1a5c31956362a6f9e0aeb3ac9c1d37b5180b0cd6d75f6d7"
   name = "github.com/grpc-ecosystem/grpc-gateway"
   packages = [
     "internal",
+    "protoc-gen-swagger/options",
     "runtime",
     "utilities",
   ]
@@ -143,14 +144,14 @@
 
 [[projects]]
   branch = "PMM-3578-obj-details"
-  digest = "1:eb95a284c900fe17b17744685ecfb55449db635821c9624ba15a8c386bbdefbf"
+  digest = "1:7cffd15c60290369c4b880a1d94ed840727c3411592500ecb73ca42146809f8f"
   name = "github.com/percona/pmm"
   packages = [
     "api/qanpb",
     "version",
   ]
   pruneopts = "NUT"
-  revision = "575652ecc8bf06b4f81494ba108f7beb2e3958f0"
+  revision = "c8f4343dae86d21038986ef07461d8e11de9fba0"
 
 [[projects]]
   digest = "1:14715f705ff5dfe0ffd6571d7d201dd8e921030f8070321a79380d8ca4ec1a24"
