+++ models/data_ingestion.go
@@ -42,6 +42,7 @@ const insertSQL = `
     metrics_source,
     period_start,
     period_length,
+    fingerprint,
     example,
     example_format,
     is_truncated,
@@ -203,6 +204,7 @@ const insertSQL = `
     CAST( :metrics_source_s AS Enum8('METRICS_SOURCE_INVALID' = 0, 'MYSQL_SLOWLOG' = 1, 'MYSQL_PERFSCHEMA' = 2)) AS metrics_source,
     :period_start_ts,
     :period_length_secs,
+    :fingerprint,
     :example,
     CAST( :example_format_s AS Enum8('EXAMPLE' = 0, 'FINGERPRINT' = 1)) AS example_format,
     :is_query_truncated,
